+++ src/loaders/common/loader_helper.py
@@ -9,13 +9,23 @@ import time
 import uuid
 from collections import defaultdict
 from contextlib import closing
+from datetime import datetime
 from pathlib import Path
 from typing import Any
 
 import jsonlines
 
 import src.common.storage.collection_and_field_names as names
-from src.common.storage.db_doc_conversions import collection_data_id_key
+from src.common.collection_column_specs.load_specs import load_spec
+from src.common.product_models.columnar_attribs_common_models import (
+    ColumnType,
+    AttributesColumn,
+    ColumnarAttributesMeta,
+)
+from src.common.storage.db_doc_conversions import (
+    collection_data_id_key,
+    collection_load_version_key,
+)
 from src.loaders.common.loader_common_names import (
     COLLECTION_SOURCE_DIR,
     DOCKER_HOST,
@@ -33,29 +43,116 @@ from src.loaders.common.loader_common_names import (
 """
 This module contains helper functions used for loaders (e.g. compute_genome_attribs, gtdb_genome_attribs_loader, etc.)
 """
-from pydantic import BaseModel
+
+NONE_STR = ['N/A', 'NA', 'None', 'none', 'null', 'Null', 'NULL', '']
 
 
-def model_to_dict(model_instance) -> dict:
+def _convert_to_iso8601(date_string: str) -> str:
+    # Convert a date string to ISO 8601 format
+    formats_to_try = ["%Y/%m/%d", "%Y-%m-%d"]  # Add more formats as needed
+
+    for date_format in formats_to_try:
+        try:
+            parsed_date = datetime.strptime(date_string, date_format)
+            iso8601_date = parsed_date.date().isoformat()
+            return iso8601_date
+        except ValueError:
+            continue
+
+    raise ValueError("Unrecognized date format")
+
+
+def _convert_values_to_type(docs: list[dict], key: str, col_type: ColumnType):
+    # Convert the values of a column to the specified type
+    values = []
+    for doc in docs:
+        try:
+            value = doc[key]
+            value = None if value in NONE_STR else value
+            if value is None:
+                doc[key] = value
+                continue
+
+            if col_type == ColumnType.INT:
+                doc[key] = int(value)
+            elif col_type == ColumnType.FLOAT:
+                doc[key] = float(value)
+            elif col_type == ColumnType.STRING:
+                doc[key] = str(value)
+            elif col_type == ColumnType.DATE:
+                doc[key] = _convert_to_iso8601(value)
+            else:
+                raise ValueError(f'casting not implemented for {col_type}')
+
+            values.append(doc[key])
+        except KeyError as e:
+            raise ValueError(f'Unable to find key: {key} in {doc}') from e
+        except ValueError as e:
+            raise ValueError(f'Unable to convert value: {key} from {doc} to type: {col_type}') from e
+
+    return values
+
+
+def process_columnar_meta(
+        docs: list[dict],
+        kbase_collection: str,
+        load_ver: str,
+        allow_missing_metadata: bool = False,
+):
     """
-    Converts a pydantic model instance to a dictionary.
+    Process the columnar metadata for the genome attributes.
 
-    :param model_instance: a pydantic model instance
+    :param docs: the list of documents
+    :param kbase_collection: the KBase collection name
+    :param load_ver: the load version
+    :param allow_missing_metadata: whether to allow missing metadata, default is False
     """
-    result = {}
 
-    for field_name, field_value in model_instance.__dict__.items():
-        if isinstance(field_value, BaseModel):
-            # If the attribute is a BaseModel, recursively convert it to a dictionary
-            result[field_name] = model_to_dict(field_value)
-        elif isinstance(field_value, list):
-            # If the attribute is a list, check each item for BaseModel instances
-            result[field_name] = [model_to_dict(item) if isinstance(item, BaseModel) else item for item in field_value]
-        else:
-            # Otherwise, include the attribute as is
-            result[field_name] = field_value
+    spec = load_spec(names.GENOME_ATTRIBS_PRODUCT_ID, kbase_collection)
+    columns = list()
+    for col_spec in spec.columns:
+        key, col_type = col_spec.key, col_spec.type
+        if key not in docs[0]:
+            error_message = f'Unable to find column: {key} in the docs'
+            if allow_missing_metadata:
+                # some genome attribute tools may not have been executed for specific collection
+                print(error_message)
+                continue
+            else:
+                raise ValueError(error_message)
 
-    return result
+        values = _convert_values_to_type(docs, key, col_type)
+        min_value, max_value, enum_values = None, None, None
+        if col_type in [ColumnType.INT, ColumnType.FLOAT, ColumnType.DATE]:
+            if values:
+                min_value, max_value = min(values), max(values)
+            else:
+                # AttributesColumn validates min_value and max_value, so we need to set them to 0 if column is empty
+                # TODO raise error instead of setting to 0?
+                min_value = 0 if col_type == ColumnType.INT else 0.0
+                max_value = 0 if col_type == ColumnType.INT else 0.0
+        elif col_type == ColumnType.ENUM:
+            enum_values = list(set(values))
+            enum_values.sort()
+
+        attri_column = AttributesColumn(
+            **col_spec.model_dump(),
+            min_value=min_value,
+            max_value=max_value,
+            enum_values=enum_values
+        )
+        columns.append(attri_column)
+
+    columnar_attri_meta = ColumnarAttributesMeta(columns=columns)
+
+    meta_doc = columnar_attri_meta.model_dump()
+    meta_doc.update({
+        names.FLD_ARANGO_KEY: collection_load_version_key(kbase_collection, load_ver),
+        names.FLD_COLLECTION_ID: kbase_collection,
+        names.FLD_LOAD_VERSION: load_ver
+    })
+
+    return docs, meta_doc
 
 
 def convert_to_json(docs, outfile):
